backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
semantic-kernel>=0.4.0
python-dotenv>=1.0.0
pyodbc>=4.0.39
pandas>=2.1.1
nest-asyncio>=1.5.8
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.13.0
python-docx>=0.8.11
markdown>=3.5.1
google-generativeai>=0.3.0